Refactor banner code for package promotions
These changes include:

Added role="region" to the section, which helps assistive technology to understand the layout and to navigate sections.
Added aria-label="Banner Promotion" to the section. This provides a name to the section which can be read by screen readers.
Added aria-label="Explore More Promotions" to the button. This provides a descriptive
name to the button.
Wrap the h4, h2, and button in a div and position that div absolutely in relation to the banner section.
.banner-content is positioned in the center of #banner with text also centered.