Refactor New Arrivals Lebanese Collection Just like before, I added role="list" to the product container and role= "listitem" to the individual product. This is to clearly indicate the structure of the items to assistive technologies.
I added an alt text to the image to describe the product in the image. However, since the information is already described in text, I added role="presentation" to indicate to screen readers that this image is decorative.

For the star rating, I added aria-label="5 out of 5 stars" to convey the star rating information to screen readers.

For the add to cart link, I added aria-label="Add Long Bread Sticks to cart" to give more specific instructions about what the link will do when activated.# with '#' will be ignored, and an empty message aborts the commit.
modified:   index.html