Refactor Newsletter You can add a <form> tag around your input and button to group them together. This gives the user a hint that these two elements interact together and is particularly important for accessibility.
Accessibility: For the input field, add a label (even if it's visually hidden), and always include a name attribute.

BEM Methodology: For CSS, a more modern way to style elements is to use the BEM methodology, which stands for Block, Element, Modifier. This methodology makes it easier to understand the relationship between HTML and CSS.# with '#' will be ignored, and an empty message aborts the commit.
This refactoring includes more semantic markup, a better focus on accessibility, and CSS that follows the BEM methodology. Also, note that I've used rem units instead of pixels to make the website more responsive and accessible.
	modified:   index.html
	modified:   style.css